+++ src/main/kotlin/icu/windea/pls/lang/intentions/script/EvaluateInlineMathIntention.kt
@@ -8,14 +8,14 @@ import com.intellij.openapi.application.EDT
 import com.intellij.openapi.project.DumbAware
 import icu.windea.pls.PlsBundle
 import icu.windea.pls.PlsFacade
-import icu.windea.pls.lang.ui.calculators.ParadoxInlineMathCalculatorDialog
+import icu.windea.pls.lang.ui.evaluators.ParadoxInlineMathEvaluatorDialog
 import icu.windea.pls.script.psi.ParadoxScriptInlineMath
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class CalculateInlineMathIntention : PsiBasedModCommandAction<ParadoxScriptInlineMath>(ParadoxScriptInlineMath::class.java), DumbAware {
-    override fun getFamilyName() = PlsBundle.message("intention.calculateInlineMath")
+class EvaluateInlineMathIntention : PsiBasedModCommandAction<ParadoxScriptInlineMath>(ParadoxScriptInlineMath::class.java), DumbAware {
+    override fun getFamilyName() = PlsBundle.message("intention.evaluateInlineMath")
 
     override fun perform(context: ActionContext, element: ParadoxScriptInlineMath): ModCommand {
         if (element.expression.isEmpty()) return ModCommand.nop()
@@ -23,7 +23,7 @@ class CalculateInlineMathIntention : PsiBasedModCommandAction<ParadoxScriptInlin
         val coroutineScope = PlsFacade.getCoroutineScope(project)
         coroutineScope.launch {
             withContext(Dispatchers.EDT) {
-                val dialog = ParadoxInlineMathCalculatorDialog(project, element)
+                val dialog = ParadoxInlineMathEvaluatorDialog(project, element)
                 dialog.show()
             }
         }
@@ -31,6 +31,6 @@ class CalculateInlineMathIntention : PsiBasedModCommandAction<ParadoxScriptInlin
     }
 
     override fun generatePreview(context: ActionContext?, element: ParadoxScriptInlineMath?): IntentionPreviewInfo {
-        return IntentionPreviewInfo.Html(PlsBundle.message("intention.calculateInlineMath.desc"))
+        return IntentionPreviewInfo.Html(PlsBundle.message("intention.evaluateInlineMath.desc"))
     }
 }